backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-parler==2.3
djangorestframework-parler==2.0
django-tinymce==3.7.0
psycopg[binary]==3.1.18
Pillow==10.2.0
python-decouple==3.8

# Development
django-debug-toolbar==4.3.0
black==24.2.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0 